requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pydantic>=2.6
pydantic-settings>=2.2
numpy>=1.24
pillow>=10.0

# Vector DB
qdrant-client>=1.9.2

# ML
transformers>=4.40,<5
safetensors>=0.4
timm>=0.9

# Detector
ultralytics>=8.2.0
opencv-python-headless>=4.8.0
